# src/main/resources/META-INF/openapi.yaml
openapi: 3.1.0
info:
  title: Ecuador Tax Document Service - Invoice Draft API
  version: 1.0.0
  description: >-
    Internal target API for creating and immediately reviewing an Invoice Draft. It performs no
    Company lookup, authentication, authorization, fiscal issuance, or SRI operation.
servers:
  - url: /api/v1
...
paths:
  /invoice-drafts/{invoiceDraftId}/fiscal-preparation:
    post:
      tags: [Fiscal Preparations]
      operationId: prepareInvoiceForFiscalIssuance
      summary: Prepare an existing Invoice Draft exactly once
      description: >-
        Accepts no request body and no caller-generated idempotency key. Natural equivalence is
        exactly the canonical X-Company-Id plus invoiceDraftId. On a first request, the service
        fixes one America/Guayaquil civil date at initial request entry, resolves authoritative
        fiscal context through the approved read-only boundary, then atomically commits the
        immutable snapshot, next controlled sequential, Numeric Code, Access Key, and baseline
        advancement. A committed replay is returned before date revalidation or external access.
        Invalid correlation input is replaced with a safe generated value and never changes
        equivalence. Any non-empty body is non-contract input and returns INVALID_REQUEST.
      parameters:
        - $ref: '#/components/parameters/FiscalPreparationInvoiceDraftId'
        - $ref: '#/components/parameters/FiscalPreparationCompanyContext'
        - $ref: '#/components/parameters/FiscalPreparationCorrelationId'
      responses:
        '201':
          description: One new Fiscal Preparation was committed.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/FiscalPreparationCorrelationId'
            Fiscal-Preparation-Replayed:
              description: False because this response committed the preparation.
              schema:
                type: boolean
                const: false
            Cache-Control:
              $ref: '#/components/headers/FiscalPreparationNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FiscalPreparationResponse'
        '200':
          description: >-
            The original committed Fiscal Preparation was returned without fiscal-context access,
            date revalidation, baseline access, or new identity generation.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/FiscalPreparationCorrelationId'
            Fiscal-Preparation-Replayed:
              description: True because this response returned the committed winner.
              schema:
                type: boolean
                const: true
            Cache-Control:
              $ref: '#/components/headers/FiscalPreparationNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FiscalPreparationResponse'
        '400':
          $ref: '#/components/responses/FiscalPreparationBadRequest'
        '404':
          $ref: '#/components/responses/FiscalPreparationDraftNotFound'
        '409':
          $ref: '#/components/responses/FiscalPreparationConflict'
        '422':
          $ref: '#/components/responses/FiscalPreparationUnprocessable'
        '500':
          $ref: '#/components/responses/FiscalPreparationInternalFailure'
        '503':
          $ref: '#/components/responses/FiscalPreparationUnavailable'
        '504':
          $ref: '#/components/responses/FiscalPreparationTimeout'
  /invoice-drafts:
    post:
      tags: [Invoice Drafts]
      operationId: createInvoiceDraft
      summary: Create and return a complete Invoice Draft
      description: >-
        Accepts authoritative Company ownership context only through X-Company-Id. Company
        identifiers are forbidden in request bodies and input schemas and are not path/query
        inputs; canonical companyId appears in the response because this feature contract
        explicitly requires it. The request contains client-controlled commercial inputs only;
        Company, Issuer, fiscal-snapshot, and system-calculated properties are prohibited. The
        service captures one requestCreationInstant at the initial request boundary before body
        consumption and derives the expected emission date from that instant in America/Guayaquil;
        that expected date remains fixed even when body consumption, validation, or commit crosses
        midnight. The upstream billing workflow
        selects the published IVA rule appropriate to the product or transaction; this service
        does not classify products or determine legal eligibility. Observable processing precedence
        is payload size, Company context, correlation, idempotency-key presence/cardinality/
        normalization/grammar, representation, Application-owned Stage 6 emission-point
        normalization/validation followed by business-text normalization, local Company-scoped
        binding lookup, equivalent replay, idempotency
        conflict, Stage 10 independent business validation, Stage 11A deterministic monetary
        calculation, the exact Stage 11B order (range/overflow; discount-over-gross;
        final-consumer calculated-total limit; total-dependent payment shape/positivity; exact
        payment reconciliation), and atomic persistence. The API adapter alone races the
        application Uni against one monotonic 10-second deadline, accepts exactly one terminal
        result, discards every late application/database outcome, and then selects/maps HTTP.
        Timeout wins when expiry becomes terminal before the application result is accepted; only
        the API maps REQUEST_TIMEOUT to 504 or any other transport-neutral failure to HTTP. An
        unresolved commit is reconciled by retrying the same CompanyId, idempotency key, and
        equivalent content.
      parameters:
        - $ref: '#/components/parameters/CompanyContext'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvoiceDraftRequest'
      responses:
        '201':
          description: >-
            A new Invoice Draft and idempotency binding were committed. The Application-created,
            timestamp-free InvoiceDraftCandidate was accepted by persistence; the persistence port
            returned an internal PersistedInvoiceDraft, which the API mapped to the declared
            InvoiceDraftResponse. Neither internal model is an HTTP representation. The response's
            createdAt and updatedAt are the exact same Instant obtained from one
            transactional-clock invocation.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
            Idempotency-Replayed:
              $ref: '#/components/headers/NewResult'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceDraftResponse'
        '200':
          description: >-
            An equivalent request returned the original committed Invoice Draft without
            revalidating its emission date against the current Ecuadorian date or invoking the
            transactional clock. The originally persisted createdAt and updatedAt are returned
            unchanged.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
            Idempotency-Replayed:
              $ref: '#/components/headers/ReplayedResult'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceDraftResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '422':
          $ref: '#/components/responses/ValidationFailure'
        '500':
          $ref: '#/components/responses/InternalFailure'
        '503':
          $ref: '#/components/responses/PersistenceUnavailable'
        '504':
          $ref: '#/components/responses/RequestTimeout'
components:
  parameters:
    FiscalPreparationInvoiceDraftId:
      name: invoiceDraftId
      in: path
      required: true
      description: Existing non-nil Invoice Draft UUID interpreted only inside X-Company-Id scope.
      schema:
        $ref: '#/components/schemas/FiscalPreparationNonNilUuid'
    FiscalPreparationCompanyContext:
      name: X-Company-Id
      in: header
      required: true
      description: >-
        Exactly one field value is accepted. Surrounding ASCII SP/HTAB is trimmed once and the
        result must be a syntactically valid non-nil UUID. It is normalized to lowercase
        hyphenated form and scopes every owned read and mutation. It is business context, not an
        authentication or authorization credential, and is absent from body, path, query, and
        response schemas.
      schema:
        $ref: '#/components/schemas/FiscalPreparationNonNilUuid'
    FiscalPreparationCorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      description: >-
        One safe value may be preserved. Missing, repeated, blank, unsafe, or over-length input is
        never echoed and receives a generated UUID replacement. Correlation never participates in
        natural equivalence.
      schema:
        type: string
        maxLength: 64
    CompanyContext:
      name: X-Company-Id
      in: header
      required: true
      description: >-
        Exactly one value is accepted. Surrounding ASCII SP/HTAB is trimmed once; the result is a
        syntactically valid non-nil UUID. The service normalizes it to canonical lowercase
        hyphenated form. The value identifies the Invoice Draft ownership and idempotency
        partition. It is business context, not an authentication or authorization credential.
      schema:
        type: string
        format: uuid
        pattern: '^(?!00000000-0000-0000-0000-000000000000$)[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}$'
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: >-
        Mandatory single-valued caller-generated request header. Exactly one field value must
        remain after HTTP header parsing. The API trims leading and trailing ASCII SP (U+0020) and
        HTAB (U+0009) once, changes no internal character or case, validates the normalized value,
        and uses that identical value for lookup, hashing, and persistence in the canonical
        CompanyId scope. Missing returns IDEMPOTENCY_KEY_REQUIRED. A single blank,
        SP/HTAB-only after that trim, over-128, control, non-ASCII, or otherwise grammar-invalid value returns
        IDEMPOTENCY_KEY_INVALID. Repeated fields, a parser result containing multiple values, or
        any comma-containing/comma-combined ambiguous value returns IDEMPOTENCY_KEY_MULTIPLE; the
        service never chooses the first value. The schema pattern describes the normalized value,
        permits internal ASCII spaces, and excludes comma. A valid binding remains for the
        lifetime of the created draft.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[\x21-\x2B\x2D-\x7E](?:[\x20-\x2B\x2D-\x7E]{0,126}[\x21-\x2B\x2D-\x7E])?$'
        examples:
          - invoice-draft-e2e-001
          - Draft Key 42
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      description: >-
        Optional caller correlation identifier. When supplied, exactly one value is accepted and
        surrounding ASCII SP/HTAB is trimmed once. A valid value contains 1 to 64 ASCII characters, begins with
        an ASCII letter or digit, and contains only ASCII letters, digits, period, underscore,
        colon, or hyphen. A valid supplied value is preserved and returned. When absent, the
        service generates and returns a safe UUID. A blank, repeated, over-length, or otherwise
        unsafe supplied value is never echoed: the service generates a replacement UUID and
        returns INVALID_REQUEST when correlation validation is the governing failure. If payload
        size is already conclusively over limit, classification only chooses the safe correlation
        returned with REQUEST_PAYLOAD_TOO_LARGE; invalid input does not replace 413 with 400.
      schema:
        $ref: '#/components/schemas/CorrelationIdentifier'
  headers:
    FiscalPreparationCorrelationId:
      description: Safe preserved or generated correlation identifier.
      schema:
        type: string
        minLength: 1
        maxLength: 64
    FiscalPreparationNoStore:
      description: Prevents intermediary caching of the sensitive successful representation.
      schema:
        type: string
        const: no-store
    CorrelationId:
      description: Correlates this attempt and its safe operational evidence.
      required: true
      schema:
        $ref: '#/components/schemas/CorrelationIdentifier'
    NewResult:
      description: Indicates that this response represents a new committed draft.
      required: true
      schema:
        type: boolean
        const: false
    ReplayedResult:
      description: >-
        Indicates that this response represents an equivalent replay. Both originally persisted
        timestamps are returned unchanged and the transactional clock is not invoked.
      required: true
      schema:
        type: boolean
        const: true
  schemas:
...
    FiscalPreparationResponse:
      type: object
      additionalProperties: false
      required:
        - fiscalPreparationId
        - invoiceDraftId
        - emissionDate
        - fiscalContextSnapshot
        - officialSequentialNumber
        - numericCode
        - accessKey
        - createdAt
      properties:
        fiscalPreparationId:
          $ref: '#/components/schemas/FiscalPreparationNonNilUuid'
        invoiceDraftId:
          $ref: '#/components/schemas/FiscalPreparationNonNilUuid'
        emissionDate:
          type: string
          format: date
          description: Exact unchanged Invoice Draft emission date.
        fiscalContextSnapshot:
          $ref: '#/components/schemas/FiscalContextSnapshot'
        officialSequentialNumber:
          type: string
          pattern: '^(?!000000000$)[0-9]{9}$'
        numericCode:
          type: string
          pattern: '^[0-9]{8}$'
        accessKey:
          type: string
          pattern: '^[0-9]{49}$'
          description: Complete SRI v2.33 Access Key required by this internal success contract.
        createdAt:
          type: string
          format: date-time
...
    CorrelationIdentifier:
      type: string
      minLength: 1
      maxLength: 64
      pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,63}$'
    CreateInvoiceDraftRequest:
      type: object
      additionalProperties: false
      description: >-
        Client-controlled inputs only. Company identifiers, including companyId, are forbidden in
        this request schema and are accepted only through X-Company-Id. The API decodes HTTP/JSON,
        rejects malformed representations, validates transport structure, and forwards decoded
        business and identifier values unchanged. It does not normalize NFC, trim business or
        identifier values, collapse business spaces, lowercase text, or derive canonicalName. At
        Stage 6, Application first validates and canonicalizes emissionPointId, then invokes
        BusinessTextNormalizer exactly once for each supplied applicable value and zero times for
        an absent optional value. canonicalName, createdAt, and updatedAt are server-derived and
        are not request properties.
      x-business-text-processing:
        owner: Application
        stage: 6
        invocationCardinality: exactly once for each supplied applicable value; zero for absent optional values
        wireSchemaScope: standard JSON Schema facets validate decoded JSON type and structure only
        postNormalizationRules: documented by each field's x-application-stage-6 extension
        generalTextOperationOrder:
          - NFC_NORMALIZATION
          - LEADING_TRAILING_U+0020_TRIM
          - PROHIBITED_CODE_POINT_AND_WHITESPACE_VALIDATION
          - DISPLAY_CODE_POINT_LENGTH_VALIDATION
          - CANONICAL_NAME_DERIVATION_WHEN_APPLICABLE
          - CANONICAL_NAME_CODE_POINT_LENGTH_VALIDATION_WHEN_APPLICABLE
      x-stage-5-property-classification:
        owner: API
        decodedRepresentation: one JSON object
        malformedUnsupportedOrNonObjectOutcome: INVALID_REQUEST
        inspectionPoint: before ordinary schema binding
        recognizedCalculatedPaths:
          - /taxTotals
          - /subtotalBeforeTaxes
          - /totalDiscount
          - /grandTotal
          - /lines/{i}/grossAmount
          - /lines/{i}/netAmount
          - /lines/{i}/lineTotal
          - /lines/{i}/tax
          - /lines/{i}/taxBase
          - /lines/{i}/taxAmount
          - /lines/{i}/taxCode
          - /lines/{i}/taxRate
          - /lines/{i}/officialTaxCode
          - /lines/{i}/officialPercentageCode
          - /lines/{i}/rate
        subtreePaths: ['/taxTotals', '/lines/{i}/tax']
        lineIndexPlaceholder: '{i} is any zero-based lines array position'
        matchOutcome: PROHIBITED_CALCULATED_FIELD
        matchIgnoresSuppliedValueAndType: true
        matchPrecedesOrdinaryUnknownMissingAndTypeErrors: true
        multipleMatchesProduceOneTopLevelOutcome: true
        violationsMember: omitted
        rejectedValueExposed: false
        noMatchOrdinaryUnknownOutcome: INVALID_REQUEST
      x-validation-precedence-examples:
        terminalResponseCardinality: exactly one
        lateOutcomeHandling: discarded
        pairs:
          - firstFact: RECOGNIZED_CALCULATED_PATH
            competingFact: ORDINARY_UNKNOWN_OR_PROHIBITED_PROPERTY
            preDeadlineOutcome: PROHIBITED_CALCULATED_FIELD
            deadlineFirstOutcome: REQUEST_TIMEOUT
          - firstFact: RECOGNIZED_CALCULATED_PATH
            competingFact: INVALID_EMISSION_POINT_STRING
            preDeadlineOutcome: PROHIBITED_CALCULATED_FIELD
            deadlineFirstOutcome: REQUEST_TIMEOUT
          - firstFact: RECOGNIZED_CALCULATED_PATH
            competingFact: APPLICATION_STAGE_6_CANONICAL_NAME_TOO_LONG
            preDeadlineOutcome: PROHIBITED_CALCULATED_FIELD
            deadlineFirstOutcome: REQUEST_TIMEOUT
          - firstFact: ORDINARY_UNKNOWN_OR_PROHIBITED_PROPERTY
            competingFact: INVALID_EMISSION_POINT_STRING
            preDeadlineOutcome: INVALID_REQUEST
            deadlineFirstOutcome: REQUEST_TIMEOUT
          - firstFact: ORDINARY_UNKNOWN_OR_PROHIBITED_PROPERTY
            competingFact: APPLICATION_STAGE_6_CANONICAL_NAME_TOO_LONG
            preDeadlineOutcome: INVALID_REQUEST
            deadlineFirstOutcome: REQUEST_TIMEOUT
          - firstFact: INVALID_EMISSION_POINT_STRING
            competingFact: APPLICATION_STAGE_6_CANONICAL_NAME_TOO_LONG
            preDeadlineOutcome: EMISSION_POINT_INVALID
            deadlineFirstOutcome: REQUEST_TIMEOUT
          - firstFact: APPLICATION_STAGE_6_CANONICAL_NAME_TOO_LONG
            competingFact: STAGE_10_EMAIL_INVALID
            preDeadlineOutcome: CANONICAL_NAME_TOO_LONG
            deadlineFirstOutcome: REQUEST_TIMEOUT
          - firstFact: STAGE_10_EMAIL_INVALID
            competingFact: DEADLINE
            preDeadlineOutcome: EMAIL_INVALID
            deadlineFirstOutcome: REQUEST_TIMEOUT
      required:
        - emissionPointId
        - emissionDate
        - buyer
        - lines
        - payments
      properties:
        emissionPointId:
          type: string
          description: >-
            Required opaque identifier. The wire schema accepts the decoded string unchanged so
            transport validation does not preempt the permitted Application transformation. At
            Stage 6, Application trims surrounding ASCII SP/HTAB exactly once, requires the result
            to be a non-nil UUID, and canonicalizes it to lowercase hyphenated form. A missing
            property or non-string JSON representation is Stage-5 INVALID_REQUEST. A decoded string
            that becomes blank, is malformed, or is nil is Stage-6 BUSINESS_VALIDATION_FAILED with
            safe violation EMISSION_POINT_INVALID. This service does not validate Company
            ownership, active state, Issuer, or establishment relationships.
          x-application-stage-6:
            owner: Application
            operationOrder:
              - LEADING_TRAILING_ASCII_SP_HTAB_TRIM_ONCE
              - NONBLANK_VALIDATION
              - NON_NIL_UUID_VALIDATION
              - LOWERCASE_HYPHENATED_UUID_CANONICALIZATION
            normalizedPattern: '^(?!00000000-0000-0000-0000-000000000000$)[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}$'
            normalizedFormat: uuid
            validExamples:
              - id: already-canonical
                rawValue: '123e4567-e89b-12d3-a456-426614174000'
                canonicalResponseValue: '123e4567-e89b-12d3-a456-426614174000'
              - id: uppercase
                rawValue: '123E4567-E89B-12D3-A456-426614174000'
                canonicalResponseValue: '123e4567-e89b-12d3-a456-426614174000'
              - id: surrounding-ascii-sp
                rawValue: ' 123e4567-e89b-12d3-a456-426614174000 '
                canonicalResponseValue: '123e4567-e89b-12d3-a456-426614174000'
              - id: surrounding-ascii-htab
                rawValue: "\t123E4567-E89B-12D3-A456-426614174000\t"
                canonicalResponseValue: '123e4567-e89b-12d3-a456-426614174000'
            invalidExamples:
              - id: missing-property
                representation: PROPERTY_ABSENT
                failureStage: 5
                outcome: INVALID_REQUEST
              - id: non-string-property
                representation: JSON_NON_STRING_INCLUDING_NULL
                failureStage: 5
                outcome: INVALID_REQUEST
              - id: empty-string
                rawValue: ''
                failureStage: 6
                outcome: EMISSION_POINT_INVALID
              - id: trim-to-empty-ascii-sp
                rawValue: '   '
                failureStage: 6
                outcome: EMISSION_POINT_INVALID
              - id: trim-to-empty-ascii-htab
                rawValue: "\t\t"
                failureStage: 6
                outcome: EMISSION_POINT_INVALID
              - id: malformed
                rawValue: not-a-uuid
                failureStage: 6
                outcome: EMISSION_POINT_INVALID
              - id: nil-uuid
                rawValue: '00000000-0000-0000-0000-000000000000'
                failureStage: 6
                outcome: EMISSION_POINT_INVALID
              - id: internal-ascii-sp
                rawValue: '123e4567-e89b-12d3-a456-4266141740 00'
                failureStage: 6
                outcome: EMISSION_POINT_INVALID
              - id: internal-ascii-htab
                rawValue: "123e4567-e89b-12d3-a456-4266141740\t00"
                failureStage: 6
                outcome: EMISSION_POINT_INVALID
              - id: braced
                rawValue: '{123e4567-e89b-12d3-a456-426614174000}'
                failureStage: 6
                outcome: EMISSION_POINT_INVALID
              - id: non-hyphenated
                rawValue: '123e4567e89b12d3a456426614174000'
                failureStage: 6
                outcome: EMISSION_POINT_INVALID
            failureOutcome:
              topLevelCode: BUSINESS_VALIDATION_FAILED
              httpStatus: 422
              violationCode: EMISSION_POINT_INVALID
              field: emissionPointId
              validationStage: NORMALIZATION
              rejectedValueExposed: false
        emissionDate:
          type: string
          format: date
          description: >-
            Must equal the Ecuadorian civil date derived in America/Guayaquil from the single
            requestCreationInstant captured at the initial request boundary before body
            consumption. That expected date is fixed across body consumption, midnight, and
            commit; requestCreationInstant is not a request property.
        buyer:
          $ref: '#/components/schemas/BuyerInput'
        lines:
          type: array
          minItems: 1
          maxItems: 500
          description: Order is business-significant and preserved.
          items:
            $ref: '#/components/schemas/InvoiceLineInput'
        payments:
          type: array
          minItems: 1
          maxItems: 8
          description: >-
            Order is representation-only; payment-method identity must be unique. Each method must
            exist, be active, and satisfy effectiveFrom <= emissionDate and (effectiveTo is null or
            emissionDate <= effectiveTo), inclusively, using this request's emissionDate only.
          items:
            $ref: '#/components/schemas/PaymentInput'
        additionalInformation:
          type: array
          maxItems: 15
          default: []
          description: >-
            Order is representation-only; persisted canonical names must be unique. At Stage 6,
            the Application normalizes and trims each supplied name once; within that same
            per-name invocation, canonical derivation reuses the normalized display value,
            collapses U+0020 runs, and lowercases with Java Locale.ROOT without a second NFC or
            trim. The result must contain at most 300 Unicode code points and is never truncated.
          items:
            $ref: '#/components/schemas/AdditionalInformationInput'
...
  responses:
    FiscalPreparationBadRequest:
      description: >-
        COMPANY_CONTEXT_REQUIRED, COMPANY_CONTEXT_INVALID, or INVALID_REQUEST. No Company-owned
        read or write occurs before valid Company context. Non-empty/prohibited request content and
        malformed or nil invoiceDraftId use INVALID_REQUEST.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/FiscalPreparationCorrelationId'
        Cache-Control:
          $ref: '#/components/headers/FiscalPreparationNoStore'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/FiscalPreparationProblemDetails'
    FiscalPreparationDraftNotFound:
      description: >-
        INVOICE_DRAFT_NOT_FOUND. Absent and cross-Company drafts are indistinguishable and expose no
        fiscal or commercial data.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/FiscalPreparationCorrelationId'
        Cache-Control:
          $ref: '#/components/headers/FiscalPreparationNoStore'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/FiscalPreparationProblemDetails'
    FiscalPreparationConflict:
      description: >-
        INVOICE_DRAFT_NOT_PREPARABLE, OFFICIAL_SEQUENCE_BASELINE_MISSING,
        OFFICIAL_SEQUENCE_BASELINE_INVALID, or OFFICIAL_SEQUENCE_EXHAUSTED. No baseline is
        initialized, repaired, wrapped, or advanced by the failed request.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/FiscalPreparationCorrelationId'
        Cache-Control:
          $ref: '#/components/headers/FiscalPreparationNoStore'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/FiscalPreparationProblemDetails'
    FiscalPreparationUnprocessable:
      description: >-
        EMISSION_DATE_STALE, FISCAL_CONTEXT_INVALID, FISCAL_CONTEXT_UNSUPPORTED, or
        FISCAL_CONTEXT_INCONSISTENT. Failure is conclusive before sequence allocation.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/FiscalPreparationCorrelationId'
        Cache-Control:
          $ref: '#/components/headers/FiscalPreparationNoStore'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/FiscalPreparationProblemDetails'
    FiscalPreparationInternalFailure:
      description: >-
        ACCESS_KEY_INVALID or INTERNAL_ERROR. ACCESS_KEY_INVALID commits no baseline advancement or
        preparation.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/FiscalPreparationCorrelationId'
        Cache-Control:
          $ref: '#/components/headers/FiscalPreparationNoStore'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/FiscalPreparationProblemDetails'
    FiscalPreparationUnavailable:
      description: >-
        FISCAL_CONTEXT_UNAVAILABLE, PERSISTENCE_FAILURE, or PREPARATION_OUTCOME_UNKNOWN. Unknown
        outcome makes no zero-state claim; retry exactly the same Company and Invoice Draft.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/FiscalPreparationCorrelationId'
        Retry-After:
          description: Optional safe retry delay when known.
          schema:
            type: string
        Cache-Control:
          $ref: '#/components/headers/FiscalPreparationNoStore'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/FiscalPreparationProblemDetails'
    FiscalPreparationTimeout:
      description: >-
        REQUEST_TIMEOUT. Zero local state is claimed only when commit did not begin or rollback is
        confirmed; otherwise the detail directs natural replay without asserting absence.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/FiscalPreparationCorrelationId'
        Cache-Control:
          $ref: '#/components/headers/FiscalPreparationNoStore'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/FiscalPreparationProblemDetails'
    BadRequest:
      description: >-
        COMPANY_CONTEXT_REQUIRED, COMPANY_CONTEXT_INVALID, IDEMPOTENCY_KEY_REQUIRED,
        IDEMPOTENCY_KEY_INVALID, IDEMPOTENCY_KEY_MULTIPLE, or INVALID_REQUEST. Repeated or
        comma-combined Idempotency-Key input is never accepted by selecting the first value.
        Invalid supplied correlation input is never echoed; a safe replacement UUID is returned.
        Missing or non-string emissionPointId representation also uses INVALID_REQUEST; decoded
        emission-point strings are handled by Stage 6 instead. Payload-size, Company-context, and
        correlation outcomes follow the defined precedence. No draft, child, or idempotency binding
        is created.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...